refactor models, update tests
