+++ gfw_pixetl/grids/grid.py
@@ -6,7 +6,7 @@ from pyproj import CRS, Transformer
 from rasterio.coords import BoundingBox
 
 from gfw_pixetl import get_module_logger
-from gfw_pixetl.utils.utils import AreaOfUse
+from gfw_pixetl.models.named_tuples import AreaOfUse
 
 LOGGER = get_module_logger(__name__)
