+++ gfw_pixetl/sources.py
@@ -16,7 +16,7 @@ from gfw_pixetl import get_module_logger
 from gfw_pixetl.connection import PgConn
 from gfw_pixetl.decorators import lazy_property
 from gfw_pixetl.errors import retry_if_rasterio_error
-from gfw_pixetl.models import Bounds
+from gfw_pixetl.models.types import Bounds
 from gfw_pixetl.settings import GDAL_ENV
 from gfw_pixetl.utils import get_bucket, utils
 from gfw_pixetl.utils.gdal import get_metadata
