+++ gfw_pixetl/tiles/tile.py
@@ -13,7 +13,9 @@ from gfw_pixetl import get_module_logger, utils
 from gfw_pixetl.errors import GDALError
 from gfw_pixetl.grids import Grid
 from gfw_pixetl.layers import Layer
-from gfw_pixetl.models import RGBA, ColorMapType, DstFormat, OrderedColorMap
+from gfw_pixetl.models.enums import ColorMapType, DstFormat
+from gfw_pixetl.models.pydantic import RGBA
+from gfw_pixetl.models.types import OrderedColorMap
 from gfw_pixetl.settings import GLOBALS
 from gfw_pixetl.sources import Destination, RasterSource
 from gfw_pixetl.utils.aws import get_s3_client
